# intelligent_reimbursement_system_server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/reimbursements/export": {
      "get": {
        "operationId": "ReimbursementController_exportExcel",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "报销类型 code，支持逗号分隔多个，如 purchase,travel",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_amount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_amount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "导出报销Excel",
        "tags": [
          "报销"
        ]
      }
    },
    "/api/reimbursements": {
      "get": {
        "operationId": "ReimbursementController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "报销类型 code，支持逗号分隔多个，如 purchase,travel",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_amount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_amount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取报销列表",
        "tags": [
          "报销"
        ]
      },
      "post": {
        "description": "请求体为 JSON 数组；数组每一项包含 applicant_name、category、apply_date、attachments、details。每个元素内的 details 数组中每一条写入一条报销记录。",
        "operationId": "ReimbursementController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateReimbursementDto"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "提交报销申请",
        "tags": [
          "报销"
        ]
      }
    },
    "/api/reimbursements/tree": {
      "get": {
        "operationId": "ReimbursementController_findTree",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "报销类型 code，支持逗号分隔多个，如 purchase,travel",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_amount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_amount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取报销树形列表",
        "tags": [
          "报销"
        ]
      }
    },
    "/api/reimbursements/{id}": {
      "patch": {
        "operationId": "ReimbursementController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveReimbursementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新报销状态",
        "tags": [
          "报销"
        ]
      }
    },
...
    "/api/roles": {
      "get": {
        "operationId": "RoleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有角色",
        "tags": [
          "角色"
        ]
      },
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建角色",
        "tags": [
          "角色"
        ]
      }
    },
    "/api/roles/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取单个角色",
        "tags": [
          "角色"
        ]
      },
      "put": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新角色",
        "tags": [
          "角色"
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除角色",
        "tags": [
          "角色"
        ]
      }
    },
    "/api/roles/{id}/permissions": {
      "put": {
        "operationId": "RoleController_assignPermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分配权限给角色",
        "tags": [
          "角色"
        ]
      }
    },
    "/api/roles/{id}/menus": {
      "put": {
        "operationId": "RoleController_assignMenus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分配菜单给角色",
        "tags": [
          "角色"
        ]
      }
    },
...
  },
  "info": {
    "title": "智能报销系统 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateReimbursementDto": {
        "type": "object",
        "properties": {
          "applicant_name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "description": "每条报销明细一条记录，对应前端一行表单；至少 1 条",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "apply_date": {
            "type": "string"
          }
        },
        "required": [
          "applicant_name",
          "category",
          "details",
          "apply_date"
        ]
      },
      "ApproveReimbursementDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "withdrawn"
            ]
          },
          "reject_reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "menus": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "label"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "menus": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}